Move compression-related code to hivemind.utils.compression
