+++ hivemind/client/averaging/__init__.py
@@ -26,7 +26,8 @@ from hivemind.client.averaging.matchmaking import Matchmaking, MatchmakingExcept
 from hivemind.client.averaging.group_info import GroupInfo
 from hivemind.proto import averaging_pb2, averaging_pb2_grpc, runtime_pb2
 from hivemind.utils.grpc import ChannelCache, GRPC_KEEPALIVE_OPTIONS, \
-    serialize_torch_tensor, deserialize_torch_tensor, split_for_streaming, combine_from_streaming
+    split_for_streaming, combine_from_streaming
+from hivemind.utils.compression import serialize_torch_tensor, deserialize_torch_tensor
 from hivemind.utils.asyncio import anext, achain, aiter, switch_to_uvloop
 from hivemind.utils.timed_storage import get_dht_time, ValueWithExpiration, DHTExpiration
 from hivemind.utils.serializer import MSGPackSerializer, SerializerBase
