+++ hivemind/utils/__init__.py
@@ -1,10 +1,11 @@
-from hivemind.utils.networking import *
-from hivemind.utils.nested import *
-from hivemind.utils.tensor_descr import *
-from hivemind.utils.serializer import *
-from hivemind.utils.mpfuture import *
-from hivemind.utils.threading import *
-from hivemind.utils.grpc import *
-from hivemind.utils.timed_storage import *
-from hivemind.utils.logging import get_logger
 from hivemind.utils.asyncio import *
+from hivemind.utils.compression import serialize_torch_tensor, deserialize_torch_tensor
+from hivemind.utils.grpc import *
+from hivemind.utils.logging import get_logger
+from hivemind.utils.mpfuture import *
+from hivemind.utils.nested import *
+from hivemind.utils.networking import *
+from hivemind.utils.serializer import *
+from hivemind.utils.tensor_descr import *
+from hivemind.utils.threading import *
+from hivemind.utils.timed_storage import *
